rename table_attribute module
