+++ sunpy/util/_table_attribute.py
